Working Library for LF-Library

+++ sketch_nov06a/lf-library/core/platform/lf_arduino_support.h
@@ -42,15 +42,13 @@ THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 /**
  * Time instant. Both physical and logical times are represented
  * using this typedef.
- * WARNING: If this code is used after about the year 2262,
- * then representing time as a 64-bit long long will be insufficient.
  */
-typedef int64_t _instant_t;
+typedef int32_t _instant_t;
 
 /**
  * Interval of time.
  */
-typedef int64_t _interval_t;
+typedef int32_t _interval_t;
 
 
 /**
