+++ sketch_nov06a/lf-library/core/pqueue.h
@@ -42,7 +42,7 @@
 #define PQUEUE_H
 
 /** priority data type */
-typedef unsigned long long pqueue_pri_t;
+typedef unsigned long pqueue_pri_t;
 
 /** callback functions to get/set/compare the priority of an element */
 typedef pqueue_pri_t (*pqueue_get_pri_f)(void *a);
